fix: Generate correct Alembic migration for guard scan logs
