--- backend/alembic/versions/480f53081907_add_guard_scan_logs.py
@@ -1,189 +1,0 @@
-"""Add guard scan logs
-
-Revision ID: 480f53081907
-Revises: 55a49e4b7bc8
-Create Date: 2026-05-23 13:26:55.227965
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '480f53081907'
-down_revision: Union[str, None] = '55a49e4b7bc8'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('rag_feedback',
-    sa.Column('id', sa.String(length=64), nullable=False),
-    sa.Column('question', sa.String(length=2000), nullable=True),
-    sa.Column('answer', sa.String(length=4000), nullable=True),
-    sa.Column('thumbs_up', sa.Integer(), nullable=True),
-    sa.Column('thumbs_down', sa.Integer(), nullable=True),
-    sa.Column('source_chunks', sa.JSON(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('users',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('email', sa.String(length=255), nullable=False),
-    sa.Column('hashed_password', sa.String(length=255), nullable=False),
-    sa.Column('full_name', sa.String(length=255), nullable=True),
-    sa.Column('company_name', sa.String(length=255), nullable=True),
-    sa.Column('subscription_tier', sa.Enum('FREE', 'STARTER', 'GROWTH', 'SCALE', name='subscriptiontier'), nullable=True),
-    sa.Column('stripe_customer_id', sa.String(length=255), nullable=True),
-    sa.Column('stripe_subscription_id', sa.String(length=255), nullable=True),
-    sa.Column('is_active', sa.Boolean(), nullable=True),
-    sa.Column('is_verified', sa.Boolean(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_users_email'), 'users', ['email'], unique=True)
-    op.create_index(op.f('ix_users_id'), 'users', ['id'], unique=False)
-    op.create_table('ai_systems',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('owner_id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(length=255), nullable=False),
-    sa.Column('description', sa.Text(), nullable=True),
-    sa.Column('version', sa.String(length=50), nullable=True),
-    sa.Column('use_case', sa.String(length=255), nullable=True),
-    sa.Column('sector', sa.String(length=255), nullable=True),
-    sa.Column('risk_level', sa.Enum('UNACCEPTABLE', 'HIGH', 'LIMITED', 'MINIMAL', name='risklevel'), nullable=True),
-    sa.Column('compliance_status', sa.Enum('NOT_STARTED', 'IN_PROGRESS', 'UNDER_REVIEW', 'COMPLIANT', 'NON_COMPLIANT', name='compliancestatus'), nullable=True),
-    sa.Column('compliance_score', sa.Float(), nullable=True),
-    sa.Column('questionnaire_responses', sa.JSON(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.ForeignKeyConstraint(['owner_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_ai_systems_id'), 'ai_systems', ['id'], unique=False)
-    op.create_table('guard_scan_logs',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('user_id', sa.Integer(), nullable=False),
-    sa.Column('prompt_hash', sa.String(length=64), nullable=False),
-    sa.Column('decision', sa.String(length=16), nullable=False),
-    sa.Column('confidence', sa.Float(), nullable=False),
-    sa.Column('matched_patterns', sa.JSON(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_guard_scan_logs_id'), 'guard_scan_logs', ['id'], unique=False)
-    op.create_index(op.f('ix_guard_scan_logs_user_id'), 'guard_scan_logs', ['user_id'], unique=False)
-    op.create_table('notifications',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('user_id', sa.Integer(), nullable=False),
-    sa.Column('notification_type', sa.String(length=100), nullable=False),
-    sa.Column('title', sa.String(length=255), nullable=False),
-    sa.Column('message', sa.Text(), nullable=False),
-    sa.Column('is_read', sa.Boolean(), nullable=True),
-    sa.Column('resource_type', sa.String(length=100), nullable=True),
-    sa.Column('resource_id', sa.Integer(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_notifications_id'), 'notifications', ['id'], unique=False)
-    op.create_table('ai_system_audit_logs',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('ai_system_id', sa.Integer(), nullable=False),
-    sa.Column('changed_by_id', sa.Integer(), nullable=False),
-    sa.Column('old_values', sa.JSON(), nullable=True),
-    sa.Column('new_values', sa.JSON(), nullable=True),
-    sa.Column('changed_at', sa.DateTime(), nullable=True),
-    sa.ForeignKeyConstraint(['ai_system_id'], ['ai_systems.id'], ),
-    sa.ForeignKeyConstraint(['changed_by_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_ai_system_audit_logs_id'), 'ai_system_audit_logs', ['id'], unique=False)
-    op.create_table('compliance_snapshots',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('ai_system_id', sa.Integer(), nullable=False),
-    sa.Column('compliance_score', sa.Integer(), nullable=False),
-    sa.Column('compliance_status', sa.String(length=50), nullable=False),
-    sa.Column('risk_level', sa.String(length=50), nullable=True),
-    sa.Column('snapshotted_at', sa.DateTime(), nullable=False),
-    sa.ForeignKeyConstraint(['ai_system_id'], ['ai_systems.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_compliance_snapshots_id'), 'compliance_snapshots', ['id'], unique=False)
-    op.create_table('documents',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('owner_id', sa.Integer(), nullable=False),
-    sa.Column('ai_system_id', sa.Integer(), nullable=True),
-    sa.Column('title', sa.String(length=255), nullable=False),
-    sa.Column('document_type', sa.Enum('TECHNICAL_DOCUMENTATION', 'RISK_ASSESSMENT', 'CONFORMITY_DECLARATION', 'DATA_GOVERNANCE', 'TRANSPARENCY_NOTICE', 'HUMAN_OVERSIGHT_PLAN', 'INCIDENT_REPORT', name='documenttype'), nullable=False),
-    sa.Column('status', sa.Enum('DRAFT', 'GENERATED', 'REVIEWED', 'APPROVED', 'ARCHIVED', name='documentstatus'), nullable=True),
-    sa.Column('content', sa.Text(), nullable=True),
-    sa.Column('file_path', sa.String(length=500), nullable=True),
-    sa.Column('version', sa.String(length=20), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.ForeignKeyConstraint(['ai_system_id'], ['ai_systems.id'], ),
-    sa.ForeignKeyConstraint(['owner_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_documents_id'), 'documents', ['id'], unique=False)
-    op.create_table('risk_assessments',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('ai_system_id', sa.Integer(), nullable=False),
-    sa.Column('assessment_type', sa.String(length=100), nullable=True),
-    sa.Column('risk_level', sa.Enum('UNACCEPTABLE', 'HIGH', 'LIMITED', 'MINIMAL', name='risklevel'), nullable=True),
-    sa.Column('findings', sa.JSON(), nullable=True),
-    sa.Column('recommendations', sa.JSON(), nullable=True),
-    sa.Column('overall_score', sa.Integer(), nullable=True),
-    sa.Column('data_governance_score', sa.Integer(), nullable=True),
-    sa.Column('transparency_score', sa.Integer(), nullable=True),
-    sa.Column('human_oversight_score', sa.Integer(), nullable=True),
-    sa.Column('robustness_score', sa.Integer(), nullable=True),
-    sa.Column('assessed_at', sa.DateTime(), nullable=True),
-    sa.Column('valid_until', sa.DateTime(), nullable=True),
-    sa.ForeignKeyConstraint(['ai_system_id'], ['ai_systems.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_risk_assessments_id'), 'risk_assessments', ['id'], unique=False)
-    op.create_table('document_versions',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('document_id', sa.Integer(), nullable=False),
-    sa.Column('version_number', sa.String(length=20), nullable=False),
-    sa.Column('content', sa.Text(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.ForeignKeyConstraint(['document_id'], ['documents.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_document_versions_id'), 'document_versions', ['id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_document_versions_id'), table_name='document_versions')
-    op.drop_table('document_versions')
-    op.drop_index(op.f('ix_risk_assessments_id'), table_name='risk_assessments')
-    op.drop_table('risk_assessments')
-    op.drop_index(op.f('ix_documents_id'), table_name='documents')
-    op.drop_table('documents')
-    op.drop_index(op.f('ix_compliance_snapshots_id'), table_name='compliance_snapshots')
-    op.drop_table('compliance_snapshots')
-    op.drop_index(op.f('ix_ai_system_audit_logs_id'), table_name='ai_system_audit_logs')
-    op.drop_table('ai_system_audit_logs')
-    op.drop_index(op.f('ix_notifications_id'), table_name='notifications')
-    op.drop_table('notifications')
-    op.drop_index(op.f('ix_guard_scan_logs_user_id'), table_name='guard_scan_logs')
-    op.drop_index(op.f('ix_guard_scan_logs_id'), table_name='guard_scan_logs')
-    op.drop_table('guard_scan_logs')
-    op.drop_index(op.f('ix_ai_systems_id'), table_name='ai_systems')
-    op.drop_table('ai_systems')
-    op.drop_index(op.f('ix_users_id'), table_name='users')
-    op.drop_index(op.f('ix_users_email'), table_name='users')
-    op.drop_table('users')
-    op.drop_table('rag_feedback')
-    # ### end Alembic commands ###
