+++ backend/alembic/versions/9579c6879964_add_guard_scan_logs.py
@@ -1,0 +1,44 @@
+"""add_guard_scan_logs
+
+Revision ID: 9579c6879964
+Revises: 55a49e4b7bc8
+Create Date: 2026-05-25 14:31:23.628006
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '9579c6879964'
+down_revision: Union[str, None] = '55a49e4b7bc8'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('guard_scan_logs',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('user_id', sa.Integer(), nullable=False),
+    sa.Column('prompt_hash', sa.String(length=64), nullable=False),
+    sa.Column('decision', sa.String(length=16), nullable=False),
+    sa.Column('confidence', sa.Float(), nullable=False),
+    sa.Column('matched_patterns', sa.JSON(), nullable=True),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_guard_scan_logs_id'), 'guard_scan_logs', ['id'], unique=False)
+    op.create_index(op.f('ix_guard_scan_logs_user_id'), 'guard_scan_logs', ['user_id'], unique=False)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f('ix_guard_scan_logs_user_id'), table_name='guard_scan_logs')
+    op.drop_index(op.f('ix_guard_scan_logs_id'), table_name='guard_scan_logs')
+    op.drop_table('guard_scan_logs')
+    # ### end Alembic commands ###
